Reorganize code into mutliple files
