+++ aws-codedeploy-appspec-assistant/pkg/assistant_test_strings.go
@@ -1,0 +1,4 @@
+package assistant
+
+var appSpecStrConversionError string = "The appSpecModel does not match the expected output model"
+var appSpecObjValidationError string = "The valid appSpec object threw errors during validation"
